Start addings a Graphable concept
