+++ src/gui/models/nodeGraph/instances/configuration.h
@@ -4,7 +4,7 @@
 #pragma once
 
 #include "classes/configuration.h"
-#include "gui/models/nodeGraph/nodeWrapper.h"
+#include "gui/models/nodeGraph/graphNodeContext.h"
 
 // Configurations need access to the CoreData to access all of their
 // children.
@@ -12,3 +12,5 @@ template <> struct GraphNodeContext<Configuration *>
 {
     using type = CoreData *;
 };
+
+std::string nodeName(Configuration *const &value);
