+++ src/gui/models/nodeGraph/instances/generator.cpp
@@ -10,7 +10,7 @@ template <> std::string nodeTypeName<Generator *>(Generator *const &value) { ret
 template <> std::string nodeTypeIcon<Generator *>(Generator *const &value) { return "qrc:/Dissolve/icons/generator.svg"; }
 
 // The title of the node
-template <> std::string nodeName<Generator *>(Generator *const &value) { return "Generator"; }
+std::string nodeName(Generator *const &value) { return "Generator"; }
 
 // Get a specific piece of information from a node by index
 template <> QVariant nodeData(Generator *const &value, int role)
