+++ src/gui/models/nodeGraph/instances/generatorGraphNode.h
@@ -1,7 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright (c) 2024 Team Dissolve and contributors
 
-#include "gui/models/nodeGraph/generatorGraphModel.h"
-#include "gui/models/nodeGraph/nodeWrapper.h"
-
 #pragma once
+
+#include "gui/models/nodeGraph/generatorGraphModel.h"
