+++ src/gui/models/nodeGraph/instances/generatorNode.h
@@ -3,7 +3,11 @@
 
 #pragma once
 
-#include "gui/models/nodeGraph/generatorGraphModel.h"
+#include "gui/models/nodeGraph/graphNodeContext.h"
+#include "classes/coreData.h"
+
+class GeneratorNode;
+
 
 // GeneratorNodes need access to the CoreData to access all of their
 // children.
@@ -11,3 +15,5 @@ template <> struct GraphNodeContext<GeneratorNode *>
 {
     using type = CoreData *;
 };
+
+std::string nodeName(GeneratorNode *const &value);
